Support worker offscreencanvas
