+++ sample/index.html
@@ -41,6 +41,15 @@
     <canvas id="renderCanvas"></canvas>
 
     <script>
+
+         var worker = window.spectorWorker = new Worker("js/worker.js");
+         const htmlCanvas = document.getElementById("renderCanvas");
+         htmlCanvas.width = document.body.clientWidth;
+         htmlCanvas.height = document.body.clientHeight;
+         const offscreen = htmlCanvas.transferControlToOffscreen();
+
+         worker.postMessage({ cmd: "start", id: "renderCanvas", canvas: offscreen }, [offscreen]);
+
         // Allow querystring to navigate easily in debug in local samples.
         var indexjs = 'js/';
         var sampleSearch = /sample=([0-9-zA-z]+)/i;
@@ -60,7 +69,7 @@
             .onReady(function () {
                 console.log('All loaded.');
             })
-            .load(['assets/js/injectSpector.js', indexjs]);
+            .load(['assets/js/injectSpector.js']);
     </script>
 </body>
