+++ src/embeddedFrontend/resultView/visualStateList/visualStateListItemComponent.ts
@@ -29,14 +29,22 @@ export class VisualStateListItemComponent extends BaseComponent<IVisualStateItem
 
         if (state.VisualState.Attachments) {
             for (const imageState of state.VisualState.Attachments) {
-                if (!imageState.src) {
+                debugger;
+                if (!imageState.src && !imageState.srcBuffer) {
                     continue;
                 }
-
-                const img = document.createElement("img");
-                img.src = encodeURI(imageState.src);
-                liHolder.appendChild(img);
-
+                if (imageState.src) {
+                    const img = document.createElement("img");
+                    img.src = encodeURI(imageState.src);
+                    liHolder.appendChild(img);
+                } else {
+                    const img = document.createElement("img");
+                    const canvas = document.createElement("canvas");
+                    const ctx = canvas.getContext("bitmaprenderer");
+                    ctx.transferFromImageBitmap(imageState.srcBuffer);
+                    img.src = canvas.toDataURL();
+                    liHolder.appendChild(img);
+                }
                 if (state.VisualState.Attachments.length > 1) {
                     const attachment = document.createElement("span");
                     attachment.innerText = imageState.attachmentName;
